requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
pydantic==2.6.1
supabase==1.2.0
httpx==0.24.1
python-dotenv==1.0.1
openai==1.12.0
google-generativeai==0.3.2
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
openai-whisper==20231117
torch==2.2.0
ffmpeg-python==0.2.0
numpy==1.26.3 